Make mock stores their own files
