+++ LoopTests/Mock Stores/MockGlucoseStore.swift
@@ -1,0 +1,44 @@
+//
+//  MockGlucoseStore.swift
+//  LoopTests
+//
+//  Created by Anna Quinlan on 8/7/20.
+//  Copyright © 2020 LoopKit Authors. All rights reserved.
+//
+
+import HealthKit
+import LoopKit
+@testable import Loop
+
+class MockGlucoseStore: GlucoseStoreProtocol {
+    func getRecentMomentumEffect(_ completion: @escaping (_ effects: [GlucoseEffect]) -> Void) {
+        let fixture: [JSONDictionary] = loadFixture("momentum_effect_bouncing")
+        let dateFormatter = ISO8601DateFormatter.localTimeDate()
+
+        return completion(fixture.map {
+            return GlucoseEffect(startDate: dateFormatter.date(from: $0["date"] as! String)!, quantity: HKQuantity(unit: HKUnit(from: $0["unit"] as! String), doubleValue: $0["amount"] as! Double))
+            }
+        )
+    }
+    
+    func getCounteractionEffects(start: Date, end: Date? = nil, to effects: [GlucoseEffect], _ completion: @escaping (_ effects: [GlucoseEffectVelocity]) -> Void) {
+        let fixture: [JSONDictionary] = loadFixture("counteraction_effect_falling_glucose")
+        let dateFormatter = ISO8601DateFormatter.localTimeDate()
+
+        return completion(fixture.map {
+            return GlucoseEffectVelocity(startDate: dateFormatter.date(from: $0["startDate"] as! String)!, endDate: dateFormatter.date(from: $0["endDate"] as! String)!, quantity: HKQuantity(unit: HKUnit(from: $0["unit"] as! String), doubleValue:$0["value"] as! Double))
+        })
+    }
+}
+
+extension MockGlucoseStore {
+    public var bundle: Bundle {
+        return Bundle(for: type(of: self))
+    }
+
+    public func loadFixture<T>(_ resourceName: String) -> T {
+        let path = bundle.path(forResource: resourceName, ofType: "json")!
+        return try! JSONSerialization.jsonObject(with: Data(contentsOf: URL(fileURLWithPath: path)), options: []) as! T
+    }
+}
+
